Treat number literals in macro correctly. <BTS:546>
git-svn-id: https://subversion.assembla.com/svn/q3/trunk@786 8af8166b-12a6-a448-a533-9086ace3f9f6

+++ q3/qm/include/qmmacro.h
@@ -392,7 +392,7 @@ public:
 	virtual ~MacroValueNumber();
 
 public:
-	void init(long n);
+	void init(unsigned int n);
 	void term();
 
 public:
